Update mainPage Desktop : make new dir for Desktop
