+++ src/app/_components/mainPage/desktop/DesktopMainLogo.tsx
