+++ src/app/_components/mainPage/mobile/MobileMainLogo.tsx
